# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Solution",
        "description": "Description of solution API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/technologies": {
            "get": {
                "tags": [
                    "Technology"
                ],
                "summary": "Get list of technologies",
                "description": "Display a listing of technologies with filtering and pagination.",
                "operationId": "270ac03d60141f2aa3425a6477500618",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search technologies by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of technologies"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Technology"
                ],
                "summary": "Create a new technology",
                "description": "Store a newly created technology.",
                "operationId": "de6ce7d4349e250082a7e1d523c87162",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "icon"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laravel"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "laravel.svg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Technology created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/technologies/{id}": {
            "delete": {
                "tags": [
                    "Technology"
                ],
                "summary": "Delete a technology",
                "description": "Remove the specified technology.",
                "operationId": "84d1fb75f463716c44eff7213c1f3d11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Technology deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Technology"
                ],
                "summary": "Update a technology",
                "description": "Update the specified technology.",
                "operationId": "15bf24b42c6926dabbdefa35fcb06852",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "icon"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Tech"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "updated-icon.svg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Technology updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/technologies/update-language/{id}": {
            "patch": {
                "tags": [
                    "Technology"
                ],
                "summary": "Sync languages for a technology",
                "operationId": "013e97e65f1f781169d99782ed087381",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "technology ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "language_ids"
                                ],
                                "properties": {
                                    "language_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Technologies updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Technologies updated successfully"
                                        },
                                        "service_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "technologies": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            },
                                            "example": [
                                                1,
                                                2,
                                                3
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/technologies/update-ittechnology/{id}": {
            "patch": {
                "tags": [
                    "Technology"
                ],
                "summary": "Sync it-technology for a technology",
                "operationId": "1037a7d8fad94b5eeb266c4bb72819f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "technology ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ittechnology_ids"
                                ],
                                "properties": {
                                    "ittechnology_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Technologies updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Technologies updated successfully"
                                        },
                                        "service_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "technologies": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            },
                                            "example": [
                                                1,
                                                2,
                                                3
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}